requirements.txt
fastapi
uvicorn
httpx
python-dotenv
langchain
langchain-openai
langchain-core
openai